Remove use of current balances from gatherCoins
Remove use of current balances from gatherCoins when logging a failure
to gather. If this information needs to be logged, it should be from the
calling code, as it's no longer in scope for this function. Additionally
the previous version read the balance even when trace level logging was
not enabled, which would have been a performance hit
e